% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFileComparator.R
\docType{class}
\name{BinaryFileComparator-class}
\alias{BinaryFileComparator-class}
\title{BinaryFileComparator.R}
\description{
Fallback comparator for binary files without any specific definied comparator.
}
\examples{

# The normal way for creating a comparator would be to call the generic factory
# method verifyr2::create_file_comparator, but if needed, an explicit comparator
# can be instantiated directly as well.
comparator <- new("BinaryFileComparator")

}
