% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryFileComparator.R
\name{BinaryFileComparator}
\alias{BinaryFileComparator}
\title{BinaryFileComparator.R}
\description{
BinaryFileComparator.R

BinaryFileComparator.R
}
\details{
Fallback comparator for binary files without any specific definied
comparator.
}
\examples{

# The normal way for creating a comparator would be to call the generic
# factory method verifyr2::create_comparator that will automatically create
# the correct comparator instance based on the file types.

file1 <- 'my_file1.bin'
file2 <- 'my_file2.bin'
comparator <- verifyr2::create_comparator(file1, file2)

# If needed, an explicit comparator can be created as well.

file1 <- 'my_file1.bin'
file2 <- 'my_file2.bin'
comparator <- BinaryFileComparator$new(file1, file2)

}
\section{Super class}{
\code{\link[verifyr2:FileComparator]{verifyr2::FileComparator}} -> \code{BinaryFileComparator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BinaryFileComparator-vrf_contents}{\code{BinaryFileComparator$vrf_contents()}}
\item \href{#method-BinaryFileComparator-vrf_contents_inner}{\code{BinaryFileComparator$vrf_contents_inner()}}
\item \href{#method-BinaryFileComparator-vrf_summary_inner}{\code{BinaryFileComparator$vrf_summary_inner()}}
\item \href{#method-BinaryFileComparator-vrf_details_inner}{\code{BinaryFileComparator$vrf_details_inner()}}
\item \href{#method-BinaryFileComparator-clone}{\code{BinaryFileComparator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="initialize"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-initialize'><code>verifyr2::FileComparator$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_details"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_details'><code>verifyr2::FileComparator$vrf_details()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_option_value"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_option_value'><code>verifyr2::FileComparator$vrf_option_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_summary"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_summary'><code>verifyr2::FileComparator$vrf_summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFileComparator-vrf_contents"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFileComparator-vrf_contents}{}}}
\subsection{Method \code{vrf_contents()}}{
Method for getting the single file contents for the comparison. This method
returns the file contents in two separate vectors inside a list. The first
vector is the file contents and the second one is the file contents with the
rows matching the omit string excluded. This method can be overwritten by
more specialized comparator classes. This method is intended to be called
only by the comparator classes in the processing and shouldn not be called
directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFileComparator$vrf_contents(file, omit, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file for which to get the contents}

\item{\code{omit}}{string pattern to omit from the comparison}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFileComparator-vrf_contents_inner"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFileComparator-vrf_contents_inner}{}}}
\subsection{Method \code{vrf_contents_inner()}}{
Method for getting the inner part for the file contents query. The method
returns the file contents in two separate vectors inside a list. The first
vector is the file contents and the second one is the file contents with the
rows matching the omit string excluded. This method can be overwritten by
more specialized comparator classes. This method is intended to be called
only by the comparator classes in the processing and shouldn not be called
directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFileComparator$vrf_contents_inner(contents, omit, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contents}}{file contents}

\item{\code{omit}}{string pattern to omit from the comparison}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFileComparator-vrf_summary_inner"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFileComparator-vrf_summary_inner}{}}}
\subsection{Method \code{vrf_summary_inner()}}{
Method for comparing the inner part for the details query. The method
returns the file contents in two separate vectors inside a list. The first
vector is the file contents and the second one is the file contents with the
rows matching the omit string excluded. This method can be overwritten by
more specialized comparator classes. This method is intended to be called
only by the comparator classes in the processing and shouldn not be called
directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFileComparator$vrf_summary_inner(omit, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omit}}{string pattern to omit from the comparison}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFileComparator-vrf_details_inner"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFileComparator-vrf_details_inner}{}}}
\subsection{Method \code{vrf_details_inner()}}{
Method for comparing the inner part for the details query. This method can be
overwritten by more specialized comparator classes. This method is intended
to be called only by the comparator classes in the processing and shouldn't
be called directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFileComparator$vrf_details_inner(omit, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omit}}{string pattern to omit from the comparison}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryFileComparator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryFileComparator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryFileComparator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
