% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{compare_files_details,FileComparator-method}
\alias{compare_files_details,FileComparator-method}
\title{Method for comparing the file details with the given comparator instance.}
\usage{
\S4method{compare_files_details}{FileComparator}(
  comparator,
  file1 = NULL,
  file2 = NULL,
  omit = NULL,
  options = NULL,
  ...
)
}
\arguments{
\item{comparator}{comparator instance used for the comparison that is meant to be created with the factory method verifyr2::create_file_comparator.}

\item{file1}{first file to compare. NULL if using same values as for comparison creation}

\item{file2}{second file to compare. NULL if using same values as for comparison creation}

\item{omit}{all lines containing the omit string will be excluded from the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}
}
\description{
Method for comparing the file details with the given comparator instance.
}
