% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{compare_files_details}
\alias{compare_files_details}
\title{Generic for comparing the file details with the given comparator instance.}
\usage{
compare_files_details(
  comparator,
  file1 = NULL,
  file2 = NULL,
  omit = NULL,
  options = NULL,
  ...
)
}
\arguments{
\item{comparator}{comparator instance used for the comparison that is meant to be created with the factory method verifyr2::create_file_comparator.}

\item{file1}{first file to compare. NULL if using same values as for comparison creation}

\item{file2}{second file to compare. NULL if using same values as for comparison creation}

\item{omit}{all lines containing the omit string will be excluded from the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}
}
\description{
Generic for comparing the file details with the given comparator instance.
}
\examples{

# file(x) in these examples are files with full file path in all the examples.

# invoking method directly with the comparator using the files stored in the comparator itself. Note that if the comprator
# was manually created without files, this will return 'no files found'.
\code{verifyr2::compare_files_details(comparator, omit = omit)}

# invoking method with explicitly given files when using single comparator instance (when handling set of specific file types for example).
\code{verifyr2::compare_files_details(comparator, file1 = file1, file2 = file2,  omit = omit)}

}
