% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdfFileComparator.R
\name{compare_files_get_contents,PdfFileComparator-method}
\alias{compare_files_get_contents,PdfFileComparator-method}
\title{Method for getting the single file contents for the comparison. This method can be overwritten by more specialized comparator classes. This
method is intended to be called only by the comparator classes in the processing and shouldn't be called directly by the user.}
\usage{
\S4method{compare_files_get_contents}{PdfFileComparator}(comparator, file, omit, options, ...)
}
\arguments{
\item{comparator}{comparator instance used for the comparison that is meant to be created with the factory method verifyr2::create_file_comparator.}

\item{file}{file for which to get the contents}

\item{omit}{all lines containing the omit string will be excluded from the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}
}
\description{
For PdfComparator, the file contents are returned based on the mode parameter if available. "text" mode is the only supported
option initally. "text" mode will return only the text content part of the PDF file.
}
