% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxtFileComparator.R
\name{compare_files_summary_inner,TxtFileComparator-method}
\alias{compare_files_summary_inner,TxtFileComparator-method}
\title{Method for comparing the inner part for the details query. This method can be
overwritten by more specialized comparator classes. This method is intended
to be called only by the comparator classes in the processing and shouldn't
be called directly by the user.}
\usage{
\S4method{compare_files_summary_inner}{TxtFileComparator}(
  comparator,
  file1,
  file2,
  omit,
  options = NULL,
  ...
)
}
\arguments{
\item{comparator}{comparator instance used for the comparison.}

\item{file1}{first file to compare}

\item{file2}{second file to compare}

\item{omit}{all lines containing the omit string will be excluded from
the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}

\item{...}{additional parameters}
}
\description{
Method for comparing the inner part for the details query. This method can be
overwritten by more specialized comparator classes. This method is intended
to be called only by the comparator classes in the processing and shouldn't
be called directly by the user.
}
