% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{create_file_comparator}
\alias{create_file_comparator}
\title{Factory method for creating comparator instance based on the given two files.}
\usage{
create_file_comparator(file1, file2, ...)
}
\arguments{
\item{file1}{first file to compare}

\item{file2}{second file to compare}
}
\description{
Factory method for creating comparator instance based on the given two files.
}
\examples{

# file(x) in these examples are files with full file path in all the examples.
\code{file1 <- paste0(fs::path_package("/extdata/base_files/file1.rtf", package = "verifyr")}
\code{file2 <- paste0(fs::path_package("/extdata/compare_files/file1.rtf", package = "verifyr")}
\code{file3 <- file1}
\code{file4 <- file2}

instantiating a new comparator instance for every comparison:
\code{verifyr2::compare_files_summary(verifyr2::create_file_comparator(file1, file2), omit = omit)}
\code{verifyr2::compare_files_details(verifyr2::create_file_comparator(file1, file2), omit = omit)}

# instantiating a comparator instance and using that same for both comparison of same files
comparator <- verifyr2::create_file_comparator(file1, file2)
\code{verifyr2::compare_files_summary(comparator, omit = omit)}
\code{verifyr2::compare_files_details(comparator, omit = omit)}

# instantiating an explicit comparator manually when comparing files of single specific type
\code{comparator <- new("RtfFileComparator")}
\code{verifyr2::compare_files_summary(comparator, file1, file2, omit = omit)}
\code{verifyr2::compare_files_summary(comparator, file3, file4, omit = omit)}

}
