% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\docType{class}
\name{FileComparator-class}
\alias{FileComparator-class}
\title{FileComparator.R}
\description{
Comparator 'base' class containing the generic comparison methods and handling for high
level checks (like file existence). This file contains also the \code{create_file_comparator}
method that is used to create the correct comparator automatically for the compared files.
}
\examples{

file(x) in these examples are files with full file path in all the examples.

instantiating a new comparator instance for every comparison:
\code{verifyr2::compare_files_summary(verifyr2::create_file_comparator(file1, file2), omit = omit)}
\code{verifyr2::compare_files_details(verifyr2::create_file_comparator(file1, file2), omit = omit)}

instantiating a comparator instance and using that same for both comparison of same files
\code{comparator <- verifyr2::create_file_comparator(file1, file2)}
\code{verifyr2::compare_files_summary(comparator, omit = omit)}
\code{verifyr2::compare_files_details(comparator, omit = omit)}

insantiating an explicit comparator manually when comparing files of single specific type
\code{comparator <- new("PdfFileComparator")}
\code{verifyr2::compare_files_summary(comparator, file1, file2, omit = omit)}
\code{verifyr2::compare_files_summary(comparator, file3, file4, omit = omit)}
\code{verifyr2::compare_files_summary(comparator, file5, file6, omit = omit)}
}
