% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{FileComparator}
\alias{FileComparator}
\title{FileComparator.R}
\description{
FileComparator.R

FileComparator.R
}
\details{
Comparator 'abstract' class containing the generic comparison methods and
handling for high level checks (like file existence). This class should never
be instantiated - doing that and calling the comparison methods will lead to
error.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file1}}{local property for storing file1}

\item{\code{file2}}{local property for storing file2}

\item{\code{file1_contents_list}}{local property for storing extracted file1 contents}

\item{\code{file2_contents_list}}{local property for storing extracted file2 contents}

\item{\code{summary_comparison}}{local property for storing summary comparison result}

\item{\code{details_comparison}}{local property for storing details comparison result}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileComparator-new}{\code{FileComparator$new()}}
\item \href{#method-FileComparator-vrf_summary}{\code{FileComparator$vrf_summary()}}
\item \href{#method-FileComparator-vrf_details}{\code{FileComparator$vrf_details()}}
\item \href{#method-FileComparator-vrf_summary_inner}{\code{FileComparator$vrf_summary_inner()}}
\item \href{#method-FileComparator-vrf_details_inner}{\code{FileComparator$vrf_details_inner()}}
\item \href{#method-FileComparator-clone}{\code{FileComparator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileComparator-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileComparator-new}{}}}
\subsection{Method \code{new()}}{
Initialize a FileComparator instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileComparator$new(file1 = NULL, file2 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file1}}{First file to compare.}

\item{\code{file2}}{Second file to compare.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileComparator-vrf_summary"></a>}}
\if{latex}{\out{\hypertarget{method-FileComparator-vrf_summary}{}}}
\subsection{Method \code{vrf_summary()}}{
Method for comparing the file summary information. This method is intended
to be implemented only this class level. For comparator specific rules,
the internal method vrf_summary_inner should be customized on lower
levels instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileComparator$vrf_summary(omit = NULL, options = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omit}}{string pattern to omit from the comparison (default = NULL)}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileComparator-vrf_details"></a>}}
\if{latex}{\out{\hypertarget{method-FileComparator-vrf_details}{}}}
\subsection{Method \code{vrf_details()}}{
Method for comparing the file details information. This method is intended
to be implemented only this class level. For comparator specific rules,
the internal method vrf_summary_inner should be customized on lower
levels instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileComparator$vrf_details(omit = NULL, options = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omit}}{string pattern to omit from the comparison (default = NULL)}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileComparator-vrf_summary_inner"></a>}}
\if{latex}{\out{\hypertarget{method-FileComparator-vrf_summary_inner}{}}}
\subsection{Method \code{vrf_summary_inner()}}{
"Abstract" method for comparing the inner part for the summary. This method
has to be overwritten by more specialized comparator classes. This method is
intended to be called only by the comparator classes in the processing and
shouldn't be called directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileComparator$vrf_summary_inner(omit, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omit}}{string pattern to omit from the comparison (default = NULL)}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileComparator-vrf_details_inner"></a>}}
\if{latex}{\out{\hypertarget{method-FileComparator-vrf_details_inner}{}}}
\subsection{Method \code{vrf_details_inner()}}{
"Abstract" method for comparing the inner part for the detailsThis method
has to be overwritten by more specialized comparator classes. This method is
intended to be called only by the comparator classes in the processing and
shouldn't be called directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileComparator$vrf_details_inner(omit, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omit}}{string pattern to omit from the comparison (default = NULL)}

\item{\code{options}}{additional comparator parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileComparator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileComparator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileComparator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
