% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nested.R
\name{get_nested}
\alias{get_nested}
\title{Get value from nested array.}
\usage{
get_nested(data, ...)
}
\arguments{
\item{data}{main level array}

\item{...}{the keys for the array path}
}
\value{
value found from the given array path or 'NA'
}
\description{
\code{get_nested} Get value from the given nested array. If the given array
path doesn't exist, returns character 'NA'.
}
\examples{

data <- list("value1" = list("value2" = list("value3" = "value4")))

# result = "value4" (found)
result <- get_nested(data, "value1", "value2", "value3")

# result = 'NA' (not found)
result <- get_nested(data, "value1", "value2", "value4")
result <- get_nested(data, "value2", "value3")

}
