% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtfFileComparator.R
\name{hex2raw}
\alias{hex2raw}
\title{Internal helper method for converting a hex string to raw vector.}
\usage{
hex2raw(hex_string)
}
\arguments{
\item{hex_string}{hexadecimal string to be converted to raw vector}
}
\description{
Internal helper method for converting a hex string to raw vector.
}
\keyword{internal}
