% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImgFileComparator.R
\docType{class}
\name{ImgFileComparator-class}
\alias{ImgFileComparator-class}
\title{ImgFileComparator.R}
\description{
Specialiced comparator for image file (jpg, jpef, png)  comparison.
This comparator contains the custom handling for handling only img content
part for the comparison.
}
\examples{

# The normal way for creating a comparator would be to call the generic
# factory method verifyr2::vrf_comparator, but if needed, an explicit
# comparator can be instantiated directly as well.
comparator <- new("ImgFileComparator")

}
