% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdfFileComparator.R
\docType{class}
\name{PdfFileComparator-class}
\alias{PdfFileComparator-class}
\title{PdfFileComparator.R}
\description{
Specialiced comparator for PDF file comparison.
This comparator contains the custom handling for handling only PDF content
part for the comparison.
}
\examples{

# The normal way for creating a comparator would be to call the generic
# factory method verifyr2::vrf_comparator, but if needed, an explicit
# comparator can be instantiated directly as well.
comparator <- new("PdfFileComparator")

}
