% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxtFileComparator.R
\docType{class}
\name{TxtFileComparator-class}
\alias{TxtFileComparator-class}
\title{TextFileComparator.R}
\description{
Fallback comparator for text files without any specific definied comparator.
This comparator contains the methods for doing basic comparisons on raw texct contents.
}
\examples{

## Not run:
# The normal way for creating a comparator would be to call the generic factory
# method verifyr2::create_file_comparator, but if needed, an explicit comparator
# can be instantiated directly as well:
comparator <- new("TxtFileComparator")
## End(Not run)

}
