% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxtFileComparator.R
\name{vrf_contents_inner,TxtFileComparator-method}
\alias{vrf_contents_inner,TxtFileComparator-method}
\title{Generic for getting the inner part for the file contents query. The method
returns the file contents in two separate vectors inside a list. The first
vector is the file contents and the second one is the file contents with the
rows matching the omit string excluded. This method can be overwritten by
more specialized comparator classes. This method is intended to be called
only by the comparator classes in the processing and shouldn't be called
directly by the user.}
\usage{
\S4method{vrf_contents_inner}{TxtFileComparator}(comparator, contents, omit, options)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}

\item{contents}{file contents}

\item{omit}{all lines containing the omit string will be excluded
from the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}
}
\description{
Generic for getting the inner part for the file contents query. The method
returns the file contents in two separate vectors inside a list. The first
vector is the file contents and the second one is the file contents with the
rows matching the omit string excluded. This method can be overwritten by
more specialized comparator classes. This method is intended to be called
only by the comparator classes in the processing and shouldn't be called
directly by the user.
}
\keyword{internal}
