% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImgFileComparator.R
\name{vrf_details_inner,ImgFileComparator-method}
\alias{vrf_details_inner,ImgFileComparator-method}
\title{Method for comparing the inner part for the details query. This method can
be overwritten by more specialized comparator classes. This method is
intended to be called only by the comparator classes in the processing and
shouldn't be called directly by the user.}
\usage{
\S4method{vrf_details_inner}{ImgFileComparator}(comparator, file1, file2, omit, options)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}

\item{file1}{first file to compare}

\item{file2}{second file to compare}

\item{omit}{all lines containing the omit string will be excluded from
the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}
}
\description{
Method for comparing the inner part for the details query. This method can
be overwritten by more specialized comparator classes. This method is
intended to be called only by the comparator classes in the processing and
shouldn't be called directly by the user.
}
\keyword{internal}
