% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImgFileComparator.R
\name{vrf_details_inner_from_files}
\alias{vrf_details_inner_from_files}
\title{Method for comparing the inner part for the details query with the file names
as the base arguments. This is a part of a group of image processing functions
that work with different image abstractions (file, image, raw image). These
methods are intended to improve the performance so that best suiting method
version can be used depending on what data is available from the earlier
method calls to the same comparator instance.}
\usage{
vrf_details_inner_from_files(comparator, file1, file2)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}

\item{file1}{first file to compare}

\item{file2}{second file to compare}
}
\description{
Method for comparing the inner part for the details query with the file names
as the base arguments. This is a part of a group of image processing functions
that work with different image abstractions (file, image, raw image). These
methods are intended to improve the performance so that best suiting method
version can be used depending on what data is available from the earlier
method calls to the same comparator instance.
}
\keyword{internal}
