% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImgFileComparator.R
\name{vrf_details_inner_from_images}
\alias{vrf_details_inner_from_images}
\title{Method for comparing the inner part for the details query with the image instances
as the base arguments. This is a part of a group of image processing functions
that work with different image abstractions (file, image, raw image). These
methods are intended to improve the performance so that best suiting method
version can be used depending on what data is available from the earlier
method calls to the same comparator instance.}
\usage{
vrf_details_inner_from_images(comparator, image1, image2)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}

\item{image1}{first image object (created with magick) to compare}

\item{image2}{second image object (created with magick) to compare}
}
\description{
Method for comparing the inner part for the details query with the image instances
as the base arguments. This is a part of a group of image processing functions
that work with different image abstractions (file, image, raw image). These
methods are intended to improve the performance so that best suiting method
version can be used depending on what data is available from the earlier
method calls to the same comparator instance.
}
\keyword{internal}
