% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{vrf_summary,FileComparator-method}
\alias{vrf_summary,FileComparator-method}
\title{Method for comparing the file summary with the given comparator instance.}
\usage{
\S4method{vrf_summary}{FileComparator}(
  comparator,
  file1 = NULL,
  file2 = NULL,
  omit = NULL,
  options = NULL
)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}

\item{file1}{first file to compare. NULL if using same values as for
comparison creation}

\item{file2}{second file to compare. NULL if using same values as for
comparison creation}

\item{omit}{all lines containing the omit string will be excluded from
the comparison (detaulf = NULL)}

\item{options}{additional comparator parameters}
}
\description{
Method for comparing the file summary with the given comparator instance.
}
