% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{vrf_supports_summary_and_full,FileComparator-method}
\alias{vrf_supports_summary_and_full,FileComparator-method}
\title{Method for getting information whether the comparator supports both summary
and full details comparisons. This method can be overwritten by
more specialized comparator classes.}
\usage{
\S4method{vrf_supports_summary_and_full}{FileComparator}(comparator)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}
}
\description{
Method for getting information whether the comparator supports both summary
and full details comparisons. This method can be overwritten by
more specialized comparator classes.
}
