% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileComparator.R
\name{vrf_supports_summary_and_full}
\alias{vrf_supports_summary_and_full}
\title{Generic for getting information whether the comparator supports both summary
and full details comparisons.}
\usage{
vrf_supports_summary_and_full(comparator)
}
\arguments{
\item{comparator}{comparator instance used for the comparison}
}
\description{
Generic for getting information whether the comparator supports both summary
and full details comparisons.
}
