% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nested.R
\name{get_nested}
\alias{get_nested}
\title{Get value from nested array.}
\usage{
get_nested(data, detault, ...)
}
\arguments{
\item{data}{main level array}

\item{...}{the keys for the array path}

\item{default}{default value returned if the array path isn't available}
}
\value{
value found from the given array path or the given default value.
}
\description{
\code{get_nested} Get value from the given nested array. If the given array
path doesn't exist, return the given default value.
}
\examples{

data <- list("value1" = list("value2" = list("value3" = "value4")))

# result = "value4" (found)
result <- get_nested(data, "value10", "value1", "value2", "value3")

# result = "value10" (not found)
result <- get_nested(data, "value10", "value1", "value2", "value4")
result <- get_nested(data, "value10", "value2", "value3")

}
