% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XlsxFileComparator.R
\name{XlsxFileComparator}
\alias{XlsxFileComparator}
\title{XlsxFileComparator.R}
\description{
XlsxFileComparator.R

XlsxFileComparator.R
}
\details{
Specialised comparator for Excel file comparison.
This comparator contains the custom handling for reading Excel file
contents for comparison.
}
\examples{

# The normal way for creating a comparator would be to call the generic
# factory method verifyr2::create_comparator that will automatically create
# the correct comparator instance based on the file types.

file1 <- 'my_file1.xlsx'
file2 <- 'my_file2.xlsx'
comparator <- verifyr2::create_comparator(file1, file2)

# If needed, an explicit comparator can be created as well.

file1 <- 'my_file1.xlsx'
file2 <- 'my_file2.xlsx'
comparator <- XlsxFileComparator$new(file1, file2)

}
\section{Super classes}{
\code{\link[verifyr2:FileComparator]{verifyr2::FileComparator}} -> \code{\link[verifyr2:BinaryFileComparator]{verifyr2::BinaryFileComparator}} -> \code{\link[verifyr2:TxtFileComparator]{verifyr2::TxtFileComparator}} -> \code{XlsxFileComparator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-XlsxFileComparator-vrf_contents}{\code{XlsxFileComparator$vrf_contents()}}
\item \href{#method-XlsxFileComparator-vrf_details_supported}{\code{XlsxFileComparator$vrf_details_supported()}}
\item \href{#method-XlsxFileComparator-clone}{\code{XlsxFileComparator$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="initialize"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-initialize'><code>verifyr2::FileComparator$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_add_debug"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_add_debug'><code>verifyr2::FileComparator$vrf_add_debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_add_debug_files"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_add_debug_files'><code>verifyr2::FileComparator$vrf_add_debug_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_close_debug"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_close_debug'><code>verifyr2::FileComparator$vrf_close_debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_details"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_details'><code>verifyr2::FileComparator$vrf_details()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_open_debug"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_open_debug'><code>verifyr2::FileComparator$vrf_open_debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_option_value"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_option_value'><code>verifyr2::FileComparator$vrf_option_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="FileComparator" data-id="vrf_summary"><a href='../../verifyr2/html/FileComparator.html#method-FileComparator-vrf_summary'><code>verifyr2::FileComparator$vrf_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="TxtFileComparator" data-id="vrf_contents_inner"><a href='../../verifyr2/html/TxtFileComparator.html#method-TxtFileComparator-vrf_contents_inner'><code>verifyr2::TxtFileComparator$vrf_contents_inner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="TxtFileComparator" data-id="vrf_details_inner"><a href='../../verifyr2/html/TxtFileComparator.html#method-TxtFileComparator-vrf_details_inner'><code>verifyr2::TxtFileComparator$vrf_details_inner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="verifyr2" data-topic="TxtFileComparator" data-id="vrf_summary_inner"><a href='../../verifyr2/html/TxtFileComparator.html#method-TxtFileComparator-vrf_summary_inner'><code>verifyr2::TxtFileComparator$vrf_summary_inner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XlsxFileComparator-vrf_contents"></a>}}
\if{latex}{\out{\hypertarget{method-XlsxFileComparator-vrf_contents}{}}}
\subsection{Method \code{vrf_contents()}}{
Method for getting the single file contents for the comparison. The
method returns the file contents in two separate vectors inside a list.
The first vector is the file contents and the second one is the file
contents with the rows matching the omit string excluded. This method
is intended to be called only by the comparator classes in the processing
and shouldn't be called directly by the user.

For XlsxFileComparator, each sheet is read and rows are converted to
tab-separated strings for text-based comparison.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XlsxFileComparator$vrf_contents(file, config, omit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file for which to get the contents}

\item{\code{config}}{configuration values}

\item{\code{omit}}{string pattern to omit from the comparison}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XlsxFileComparator-vrf_details_supported"></a>}}
\if{latex}{\out{\hypertarget{method-XlsxFileComparator-vrf_details_supported}{}}}
\subsection{Method \code{vrf_details_supported()}}{
Inherited method for indicating whether detailed comparison is available
with the current comparator. Returns an empty string if the comparator
is supported, otherwise a string that will be concatenated with the
summary string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XlsxFileComparator$vrf_details_supported(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{configuration values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-XlsxFileComparator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-XlsxFileComparator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{XlsxFileComparator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
